check creator for Expert, Req_Res, Ques_Ans

+++ src/main/java/org/vietsearch/essme/controller/RequestResponseController.java
@@ -17,6 +17,7 @@ import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/requests")
@@ -64,8 +65,15 @@ public class RequestResponseController {
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public Request updateRequest(@PathVariable("id") String id, @Valid @RequestBody Request request) {
+    public Request updateRequest(AuthenticatedRequest authenticatedRequest,@PathVariable("id") String id, @Valid @RequestBody Request request) {
+        String uuid = authenticatedRequest.getUserId();
         if (requestRepository.existsById(id)) {
+            // check
+            if(!matchUserRequest(uuid, id)){
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied", null);
+            }
+
+            // update
             request.set_id(id);
             requestRepository.save(request);
             return request;
@@ -76,8 +84,15 @@ public class RequestResponseController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public String deleteRequest(@PathVariable("id") String id) {
+    public String deleteRequest(AuthenticatedRequest authenticatedRequest,@PathVariable("id") String id) {
+        String uuid = authenticatedRequest.getUserId();
         if (requestRepository.existsById(id)) {
+            // check
+            if(!matchUserRequest(uuid, id)){
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Permission denied", null);
+            }
+
+            // delete
             requestRepository.deleteById(id);
             return "Deleted";
         } else {
@@ -104,11 +119,12 @@ public class RequestResponseController {
 
     @PutMapping("/{requestId}/responses/{responsesId}")
     @ResponseStatus(HttpStatus.OK)
-    public Response updateResponse(@PathVariable("requestId") String requestId, @PathVariable("responsesId") String responsesId, @Valid @RequestBody Response response) {
+    public Response updateResponse(AuthenticatedRequest authenticatedRequest,@PathVariable("requestId") String requestId, @PathVariable("responsesId") String responsesId, @Valid @RequestBody Response response) {
+        String uuid = authenticatedRequest.getUserId();
         Request request = requestRepository.findById(requestId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Request not found", null));
         if (request.getResponses() != null) {
             for (Response res : request.getResponses()) {
-                if (res.get_id().equals(responsesId)) {
+                if (matchUserResponse(uuid, responsesId, res)) {
                     res.setExpertId(response.getExpertId());
                     res.setContent(response.getContent());
                     res.setUpdatedAt(new Date());
@@ -122,11 +138,12 @@ public class RequestResponseController {
 
     @DeleteMapping("/{requestId}/responses/{responsesId}")
     @ResponseStatus(HttpStatus.OK)
-    public String deleteAnswer(@PathVariable("requestId") String requestId, @PathVariable("responsesId") String responseId) {
+    public String deleteResponse(AuthenticatedRequest authenticatedRequest, @PathVariable("requestId") String requestId, @PathVariable("responsesId") String responseId) {
+        String uuid = authenticatedRequest.getUserId();
         Request request = requestRepository.findById(requestId).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Request not found", null));
         if (request.getResponses() != null) {
             for (Response res : request.getResponses()) {
-                if (res.get_id().equals(responseId)) {
+                if (matchUserResponse(uuid, responseId, res)) {
                     request.getResponses().remove(res);
                     requestRepository.save(request);
                     return "Deleted";
@@ -134,6 +151,19 @@ public class RequestResponseController {
             }
         }
         throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Response not found", null);
+    }
+
+    private boolean matchUserRequest(String uuid, String requestId){
+        // return true if uuid created request
+        Optional<Request> optional = requestRepository.findById(requestId);
+        return optional.map(res -> res.getCustomerId().equals(uuid)).orElse(false);
+    }
+
+    private boolean matchUserResponse(String uuid, String responseChangedId, Response response){
+        // return true if uuid created response
+        if(!response.get_id().equals(responseChangedId))
+            return false;
+        return response.getExpertId().equals(uuid);
     }
 
 }
